Added built version of game
As of the moment, this is the latest version of the game, and is the final deliverable for my project. I may return to this project and continue development, possibly even publishing the game one day, but this will likely be the last update to this repository for at least a few weeks.